cms/pyproject.toml
[tool.poetry]
name = "cms"
version = "0.1.0"
description = ""
authors = ["Robert Tsui <rt.roberttsui@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
markdown = "^3.7"
pyyaml = "^6.0.2"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
